add_user: Add user to group with same name when not specified
If both user and group is provided to add_user, group would be
created and user will added to the group while useradd. But when
group is not provided to add_user, it used to just create user
and later would try to fetch the group (which would be empty)
Because of this KeyError exception was being raised.

This patch fixes above issue by doing following things.

    * Create group/user only if they are not present
    * Used -U option to add user to group of same name when
      group is not provided to add_user. The -U option will
      override whatever is in /etc/login.defs
    * Use -g to add user to the group when it's already present

Signed-off-by: M S Vishwanath Bhat <msvbhat@gmail.com>