Adding config values per test case and running accordingly
With this patch, each test case will be able to have its own
config values. The config should be specified in yaml format
and should be sprcified in docstring of the test case.

This config values will be used to run the same test case with
different volume configurations. This will be taken care by
distaf itself. But the test case writer has to get the name of
the volume and mount it inside the testcase. The cleanup will
also be done outside of the test case. The test case itself will
not have anything more than test case execution part.

For example

@testcase("gluster_basic_test")
def my_basic_test():
    """
        runs_on_volume: ALL
        runs_on_protocol: [ glusterfs, nfs, cifs ]
    """

This indicates that the test case should be run against all
possible type of volume configurations. Now distaf sees it
and runs this test in against all possible volume types.

This can be overridden by a global flag global_mode: True in
the config.yml.

Note: The iterating through different protocols has not been
      done yet.

Signed-off-by: M S Vishwanath Bhat <msvbhat@gmail.com>