Adding support to write the test cases as python classes
This patch if the testcase is function or not. If it is a
function, it makes a call to the function, collects the
return value and asserts Pass/Fail

If the testcase is not a function it is assumed to be python
class or subclass, it will instantiate the class. The class
is assumed to have three methods.  setup(), run(), and
cleanup(). First setup() is called and only upon its successful
completion, run() is called. The cleanup() function is always
called at the end. The test is marked as Pass, if all three
methods return True. Else the test case is marked False

Signed-off-by: M S Vishwanath Bhat <msvbhat@gmail.com>