Adding base class for all other tests to use
The DistafBaseClass is a base class which is to be used by
all test cases. This class initialises all the required
configuration variables and setup the volume for the test
cases to use.

This class has 2 methods, setup() and teardown()

The setup() methods takes care of setting up the volume. If
the test cases needs any more setup, that needs to be taken
care in the test case itself.

The teardown() methos umounts the volume. If anymore cleaning
up part is required, test cases should handle it.

Signed-off-by: M S Vishwanath Bhat <msvbhat@gmail.com>