Initialising the second level dict inside the establish_connection
Until now, the second level dictionary was being initialised before
calling the establish_connection, in __init__ of BigBang(). This
will be problem iwe try to establish connection with the machine
which is not initialised in __init__. This patch moves the
initialisation of empty second level dictionary inside the function
establish_connection.

Also made a trivial changes of formating the code block

Signed-off-by: M S Vishwanath Bhat <msvbhat@gmail.com>