Validating the peer status to be pass if there is only server
peer_status used to return an empty string if there is only one
server in the cluster. And check in validate_peer_status used to
infer that as failure. But now peer_status returns [] which is an
empty string when there is only one server in the cluster. And
returns None when failure. The checks in validate_peer_status has
also been updated to check properly for empty list or None.

This commit fixes #16

Signed-off-by: M S Vishwanath Bhat <msvbhat@gmail.com>